feat: merged input floating with text
